backend/requirements.txt
# Base Django Requirements
Django>=5.2.6,<6.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0

# Database
# SQLite is included with Django for development
# Uncomment below for PostgreSQL in production
# psycopg2>=2.9.0

# Caching and Performance
redis>=4.5.0
django-redis>=5.3.0

# Security
cryptography>=3.4.8

# Production WSGI Server
gunicorn>=20.1.0

# Environment Management
python-dotenv>=0.19.0
